requirements.txt
llama-index
openai
llama-index-vector-stores-pinecone
llama-index-embeddings-openai
llama-index.retrievers.bm25
llama-index-llms-openai
llama-index-postprocessor-cohere-rerank
llama-index-readers-google
pinecone
pymupdf
pytesseract
pdf2image
pillow
python-dotenv
pinecone
cohere
pdfplumber
unstructured[local-inference]
load_dotenv
google-api-python-client
flask
google-auth-httplib2 
google-auth-oauthlib